ai-compliance-checker/requirements.txt
flask>=2.0
python-dotenv>=1.0
python-docx>=0.8.11